fix(uppercut more accurate)
On the edge case that the player was spamming left click and the player was on the final combo attack, due to lack of condition catching, when they spammed the left click it would cycle over back to the first combo attack while simultaniously be doing the final combo attack but of course when it did collide with enemy, in this scenario, it would not register an attack at all.

this has been fixed by adding an additional condition to check if the player's gameplay tag for the ability is not there (which indicates that the ability is over) and only if both conditions are good it will Reset the combo.