fix(random crashing in middle of combat)
During combat there were cases where the game would crash. this was because to the custom UObject "StatAdjustment" was getting cleaned up the GarbageCollector to early resulting in a invalid pointer. this was resolved by marking the variable with a UPROPERTY() to prevent the garbage collector from cleaning it up when it wasnt in use for a while.